fixup! add api for easy installation
